requirements.txt
contourpy
cycler
docker-pycreds
einops
filelock
flake8==7.0.0
flake8-bugbear==24.2.6
flake8-comprehensions==3.14.0
fonttools
fsspec
fvcore==0.1.5.post20221221
gitdb
GitPython
h5py>=3.10.0
huggingface-hub>=0.22.0
idna
imageio
imath
iopath
isort
Jinja2
jmespath
kiwisolver
MarkupSafe
matplotlib
mccabe
mpmath
msgpack
mypy-extensions
ndindex
networkx
ninja
numexpr
opencv-python
packaging
pandas
pathspec
pillow==10.2.0
platformdirs
portalocker
protobuf==4.25.3
psutil
py-cpuinfo
pycodestyle
pyflakes
pyparsing
python-dateutil
pytz
PyYAML
requests
safetensors
scipy
sentry-sdk
setproctitle
six
smmap
sympy
tables
tabulate
termcolor
timm
tqdm
triton==2.2.0
typing_extensions
tzdata==2024.1
urllib3==1.26.13
wandb
yacs
diffusers==0.29.1
numpy==1.26.4
matplotlib==3.8.4
transformers==4.41.2
accelerate==0.30.1
mediapy==1.2.0
fire==0.6.0
decord==0.6.0
OpenEXR==3.2.4
omegaconf
jaxtyping
tyro
tensorboard
kornia
lpips
matplotlib
natsort
opencv_python
pandas
packaging
peft
Pillow
plotly
protobuf
pudb
pyquaternion
munch
hydra-core==1.1.0
pytorch_lightning==1.6.0
evo